Moved JWK related classes to security.jwk package.
