+++ src/main/java/org/humancellatlas/ingest/security/jwk/JwkVault.java
@@ -1,4 +1,4 @@
-package org.humancellatlas.ingest.security;
+package org.humancellatlas.ingest.security.jwk;
 
 import java.security.PublicKey;
